feat(table-improvements): Moves some files around.
